Cambio el cooldown de float a double, para que coincida con el tiempo.

+++ src/common/global_configs.cpp
@@ -49,12 +49,12 @@ GlobalConfigs::GlobalConfigs() {
   this->bullet1MaxAmmo = gameConfigs["bullet_1_max_ammo"].as<uint16_t>();
   this->bullet1Speed = gameConfigs["bullet_1_speed"].as<uint8_t>();
   this->bullet1Damage = gameConfigs["bullet_1_damage"].as<uint8_t>();
-  this->bullet1Cooldown = gameConfigs["bullet_1_cooldown"].as<float>();
+  this->bullet1Cooldown = gameConfigs["bullet_1_cooldown"].as<double>();
 
   this->bullet2MaxAmmo = gameConfigs["bullet_2_max_ammo"].as<uint16_t>();
   this->bullet2Speed = gameConfigs["bullet_2_speed"].as<uint8_t>();
   this->bullet2Damage = gameConfigs["bullet_2_damage"].as<uint8_t>();
-  this->bullet2Cooldown = gameConfigs["bullet_2_cooldown"].as<float>();
+  this->bullet2Cooldown = gameConfigs["bullet_2_cooldown"].as<double>();
 
   this->respawnTime = gameConfigs["respawn_time"].as<double>();
 }
@@ -131,7 +131,7 @@ uint8_t GlobalConfigs::getBullet1Speed() const { return this->bullet1Speed; }
 
 uint8_t GlobalConfigs::getBullet1Damage() const { return this->bullet1Damage; }
 
-float GlobalConfigs::getBullet1Cooldown() const {
+double GlobalConfigs::getBullet1Cooldown() const {
   return this->bullet1Cooldown;
 }
 
@@ -143,7 +143,7 @@ uint8_t GlobalConfigs::getBullet2Speed() const { return this->bullet2Speed; }
 
 uint8_t GlobalConfigs::getBullet2Damage() const { return this->bullet2Damage; }
 
-float GlobalConfigs::getBullet2Cooldown() const {
+double GlobalConfigs::getBullet2Cooldown() const {
   return this->bullet2Cooldown;
 }
